Parser: fix double quoted string after comment
In Parser, the field column is used to implement 6.1.3. quoting
rules[1]. More specifically to store the indentation of the last token,
so that it can be reused when stripping whitespaces inside double quoted
string tokens.

If double quoted string comes after a block comment, this indentation
was calculated incorrectly (it did not take the ending symbol '*/' into
account when column was updated.

[1]: https://www.rfc-editor.org/rfc/rfc7950.html#section-6.1.3